Move SupportItem component to shared folder
